Compute the univariate mode estiamtes via Bernstein polynomials.

Compute the mode estimate based on Bernstein polynomials density function with automated mechanism of choosing appropriate degree of Bernstein polynomials.

Based on the smooth options, this R function returns either simple empirical mode estimate, kenerl density mode estimate or Bernstein polynomials based mode estimate (default).

emp_mode(
  data,
  fix_lower = NULL,
  fix_upper = NULL,
  smooth = TRUE,
  smooth_option = "Bernstein",
  smooth_density = FALSE,
  density_points = 512L,
  m_degree = NULL,
  sample_size = 1000L,
  m_knots = NULL
)

Arguments

data
An univariate sample data.

fix_lower
A lower bound of the support; If NULL (default), set min0, the minimum of the sample data; Set a particualr value otherwise.

fix_upper
An upper bound of the support; If NULL (default), set max1, the maximum of the sample data; Set a particualr value otherwise.

smooth
An indicator of if performing Bernstein polynomials smoothing or not; If TRUE (default), adopt Bernstein polynomials density function; If FALSE, adopt empirical density function without smoothing.

smooth_option
If smooth = TRUE, an option of smoothing method for empirical density function; "Bernstein" (default) indicates Bernstein polynomials density estimate, and "kernel" the kernel density estimate.

smooth_density
If smooth = TRUE, an indicator of if output the density function or not; If FALSE (default), only the mode estimate is given; If TRUE, the smooth density function is also included in the output.

density_points
If smooth = TRUE, a number of equally spaced points for the output density function is provided or not, default is 512.

m_degree
If smooth = TRUE and smooth_option = "Bernstein", a degree of Bernstein polynomials is provided or not; If NA (default), pick the one with maximal p-value until exceeding 0.95; Set a particualr value otherwise.

sample_size
If smooth = TRUE and smooth_option = "Bernstein", a subsample size used for Anderson-Darling criterion is provided or not, and default is 1e+03.

m_knots
If smooth = TRUE and smooth_option = "kernel", a number of knots in finding the maximum of empirical density function is provided or not; If NA (default), set m_knots = sample size of data; Set a particualr value otherwise.

Value

An empirical mode of data, a degree of Bernstein polynomials adopted with the maximum p-value and the corresponding maximum p-value of the Anderson-Darling criterion based on the subsample and Bernstein polynomials; NA if not applicable.

References

https://doi.org/10.1016/j.csda.2020.107046

Author

Bowen Liu, bliu17@ncsu.edu

Examples

n = 1e6
set.seed(1)
data = rbeta(n, 2, 5)
## Return Bernstein polynomials based mode estimate
emp_mode(data)

#> 
#> maximum point: 0.200715, probability around maximum: 0.038835, probability of left boundary: 0.00354, probability of right boundary: 0
#> The mode estimate is chosen as the maximum point of Bernstein polynomials density estimate.
#> 

#> $mode
#> [1] 0.2007149
#> 
#> $optimal_alpha
#> [1] 0.35
#> 
#> $`optimal_p-value`
#> [1] 0.9963385
#> 

## Return kenerl density mode estimate
emp_mode(data, smooth_option = "kernel")

#> $mode
#> [1] 0
#> 

## Return simple empirical mode estimate
emp_mode(data, smooth = FALSE)

#> The mode estimate is choosen as the mid point of interval with maximum difference among knots equally-spaced intervals.

#> $`number of knots`
#> [1] 1000000
#> 
#> $mode
#> [1] 0.2081525
#>